데코레이터 패턴 적용
- 동적으로 기능 추가
- 한국 통화에 맞게 cost 자료형 변경

+++ src/main/java/decoratorPattern/Coffee.java
@@ -7,5 +7,5 @@ public abstract class Coffee {
         return name;
     }
 
-    public abstract double cost();
+    public abstract int cost();
 }
